utils: Move Git worktree helpers
The Git utility layer has named modules for object IO, repository identity,
and index plumbing, while worktree mutation helpers still share the command
execution wrapper.

The project lacks a named boundary for operations that apply patches to the
worktree, check out paths, remove files, reset hard, apply stashes, or
materialize submodules. Command code must import those wrappers from the
broad command module.

This commit addresses that boundary gap by moving worktree-facing wrappers
into utils.git_worktree and updating source callers to import that module
directly.

The next commit will add architecture coverage for the Git worktree helper
boundary.
