+++ src/git_stage_batch/commands/file_scope/discard_to_batch.py
@@ -24,7 +24,7 @@ from ...data.session import snapshot_files_if_untracked
 from ...exceptions import exit_with_error
 from ...i18n import _
 from ...utils.file_io import read_text_file_line_set
-from ...utils.git import (
+from ...utils.git_worktree import (
     git_apply_to_worktree,
     git_remove_paths,
 )
